# edge/src/main/resources/api/openapi.yaml
openapi: 3.0.3
info:
  title: Police Incident Management System API
  description: |
    REST API for the Police Incident Management System. This is an event-driven system where 
    all operations produce events to Kafka. Events represent requests/commands from the edge, 
    not state changes.
    
    ## Architecture
    - Edge servers receive HTTP requests and produce events to Kafka
    - Events represent requests/commands, not state changes
    - No state reconstruction in edge layer
    - All operations are asynchronous via Kafka events
    
  version: 1.0.0
  contact:
    name: Police System API Support
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api.policesystem.example.com/api/v1
    description: Production server
...
paths:
...
  /incidents:
    post:
      tags:
        - Incidents
      summary: Report a new incident
      description: Reports a new incident and produces a ReportIncidentRequested event to Kafka
      operationId: reportIncident
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportIncidentRequest'
      responses:
        '201':
          description: Incident report request created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /incidents/{incidentId}:
    put:
      tags:
        - Incidents
      summary: Update incident information
      description: Updates incident information and produces an UpdateIncidentRequested event to Kafka
      operationId: updateIncident
      parameters:
        - $ref: '#/components/parameters/IncidentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateIncidentRequest'
      responses:
        '200':
          description: Incident update request processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags:
        - Incidents
      summary: Change incident status
      description: Changes incident status and produces a ChangeIncidentStatusRequested event to Kafka
      operationId: changeIncidentStatus
      parameters:
        - $ref: '#/components/parameters/IncidentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeStatusRequest'
      responses:
        '200':
          description: Incident status change request processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /incidents/with-relations:
    post:
      tags:
        - Incidents
      summary: Create incident with relations
      description: Creates an incident and links location/calls in one call for UI convenience
      operationId: createIncidentWithRelations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIncidentWithRelationsRequest'
      responses:
        '201':
          description: Incident created with relations
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/IncidentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /incidents/{incidentId}/dispatch:
    post:
      tags:
        - Incidents
      summary: Dispatch incident
      description: Dispatches an incident and produces a DispatchIncidentRequested event to Kafka
      operationId: dispatchIncident
      parameters:
        - $ref: '#/components/parameters/IncidentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DispatchRequest'
      responses:
        '200':
          description: Incident dispatch request processed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/DispatchIncidentResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /incidents/{incidentId}/arrive:
    post:
      tags:
        - Incidents
      summary: Arrive at incident
      description: Records arrival at incident and produces an ArriveAtIncidentRequested event to Kafka
      operationId: arriveAtIncident
      parameters:
        - $ref: '#/components/parameters/IncidentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArriveRequest'
      responses:
        '200':
          description: Arrival request processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /incidents/{incidentId}/clear:
    post:
      tags:
        - Incidents
      summary: Clear incident
      description: Clears an incident and produces a ClearIncidentRequested event to Kafka
      operationId: clearIncident
      parameters:
        - $ref: '#/components/parameters/IncidentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClearRequest'
      responses:
        '200':
          description: Clear incident request processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================
  # Calls
  # ============================================
...
  /calls/{callId}/incidents:
    post:
      tags:
        - Calls
      summary: Link call to incident
      description: Links a call to an incident and produces a LinkCallToIncidentRequested event to Kafka
      operationId: linkCallToIncident
      parameters:
        - $ref: '#/components/parameters/CallId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LinkCallToIncidentRequest'
      responses:
        '200':
          description: Call link request processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /activities/{activityId}/incidents:
    post:
      tags:
        - Activities
      summary: Link activity to incident
      description: Links an activity to an incident and produces a LinkActivityToIncidentRequested event to Kafka
      operationId: linkActivityToIncident
      parameters:
        - $ref: '#/components/parameters/ActivityId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LinkActivityToIncidentRequest'
      responses:
        '200':
          description: Activity link request processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================
  # Assignments
  # ============================================
...
components:
  parameters:
...
    IncidentId:
      name: incidentId
      in: path
      required: true
      description: Incident identifier
      schema:
        type: string
        example: "INC-001"
    CallId:
      name: callId
      in: path
      required: true
      description: Call identifier
      schema:
        type: string
        example: "CALL-001"
    ActivityId:
      name: activityId
      in: path
      required: true
      description: Activity identifier
      schema:
        type: string
        example: "ACT-001"
...
  schemas:
...
    IncidentStatus:
      type: string
      enum:
        - Reported
        - Dispatched
        - Arrived
        - In-Progress
        - Cleared
        - Closed
      example: "Reported"
...
    Priority:
      type: string
      enum:
        - Emergency
        - High
        - Medium
        - Low
      example: "High"
...
    IncidentType:
      type: string
      enum:
        - Traffic
        - Theft
        - Assault
        - Domestic
        - Burglary
        - Other
      example: "Traffic"
...
    ReportIncidentRequest:
      type: object
      required:
        - incidentId
        - priority
        - status
        - incidentType
      properties:
        incidentId:
          type: string
          example: "INC-001"
        incidentNumber:
          type: string
          example: "2024-001"
        priority:
          $ref: '#/components/schemas/Priority'
        status:
          $ref: '#/components/schemas/IncidentStatus'
        reportedTime:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        description:
          type: string
          example: "Traffic accident at Main St and Oak Ave"
        incidentType:
          $ref: '#/components/schemas/IncidentType'
    UpdateIncidentRequest:
      type: object
      properties:
        priority:
          $ref: '#/components/schemas/Priority'
        description:
          type: string
          example: "Updated description"
        incidentType:
          $ref: '#/components/schemas/IncidentType'
    DispatchRequest:
      type: object
      required:
        - dispatchedTime
      properties:
        dispatchedTime:
          type: string
          format: date-time
          example: "2024-01-15T10:35:00Z"
    ArriveRequest:
      type: object
      required:
        - arrivedTime
      properties:
        arrivedTime:
          type: string
          format: date-time
          example: "2024-01-15T10:40:00Z"
    ClearRequest:
      type: object
      required:
        - clearedTime
      properties:
        clearedTime:
          type: string
          format: date-time
          example: "2024-01-15T11:00:00Z"
...
    LinkCallToIncidentRequest:
      type: object
      required:
        - incidentId
      properties:
        incidentId:
          type: string
          example: "INC-001"
...
    LinkActivityToIncidentRequest:
      type: object
      required:
        - incidentId
      properties:
        incidentId:
          type: string
          example: "INC-001"
...
    ChangeStatusRequest:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          example: "Active"

    # Response Schemas
...
    IncidentResponse:
      type: object
      properties:
        incidentId:
          type: string
          example: "INC-001"
        incidentNumber:
          type: string
          example: "2024-001"
        relatedResources:
          $ref: '#/components/schemas/RelatedResources'
          nullable: true
    DispatchIncidentResponse:
      type: object
      properties:
        incidentId:
          type: string
          example: "INC-001"
        incidentNumber:
          type: string
          example: "2024-001"
          nullable: true
        dispatchId:
          type: string
          example: "DISP-ABC12345"
          description: "Generated dispatch ID for linking operations"
        dispatchedTime:
          type: string
          format: date-time
          example: "2024-01-15T10:40:00.000+00:00"
        relatedResources:
          $ref: '#/components/schemas/RelatedResources'
          nullable: true
    RelatedResources:
      type: object
      description: Related resource IDs for UI convenience
      properties:
        dispatchId:
          type: string
          example: "DISP-001"
          nullable: true
        assignmentIds:
          type: array
          items:
            type: string
          example: ["ASSIGN-001", "ASSIGN-002"]
          nullable: true
        callIds:
          type: array
          items:
            type: string
          example: ["CALL-001"]
          nullable: true
        locationId:
          type: string
          example: "LOC-001"
          nullable: true
        activityIds:
          type: array
          items:
            type: string
          example: ["ACT-001"]
          nullable: true
        involvedPartyIds:
          type: array
          items:
            type: string
          example: ["INV-001"]
          nullable: true
        unitId:
          type: string
          example: "UNIT-001"
          nullable: true
        vehicleId:
          type: string
          example: "VEH-001"
          nullable: true
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Request processed successfully"
        data:
          description: Response data (type varies by endpoint)
        metadata:
          $ref: '#/components/schemas/ResponseMetadata'
          nullable: true
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Bad Request"
        message:
          type: string
          example: "Validation failed"
        details:
          type: array
          items:
            type: string
          example: ["badgeNumber is required"]
        validValues:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          description: Map of field names to their valid enum values (only present for enum validation errors)
          example:
            unitType: ["Single", "Team", "Squad"]
    ResponseMetadata:
      type: object
      properties:
        correlationId:
          type: string
          example: "550e8400-e29b-41d4-a716-446655440000"
          description: "Correlation ID for request tracing"
        timestamp:
          type: string
          format: date-time
          example: "2024-01-15T10:40:00.000+00:00"
        version:
          type: string
          example: "1.0"
        links:
          type: object
          additionalProperties:
            type: string
          description: "HATEOAS-style links to related resources"
          example:
            self: "/api/v1/incidents/INC-001"
            dispatch: "/api/v1/incidents/INC-001/dispatch"
    CreateIncidentWithRelationsRequest:
      type: object
      required:
        - incidentId
        - priority
        - status
        - incidentType
      properties:
        incidentId:
          type: string
          example: "INC-001"
        incidentNumber:
          type: string
          example: "2024-001"
        priority:
          $ref: '#/components/schemas/Priority'
        status:
          $ref: '#/components/schemas/IncidentStatus'
        reportedTime:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00.000+00:00"
        description:
          type: string
          example: "Traffic accident at Main St and Oak Ave"
        incidentType:
          $ref: '#/components/schemas/IncidentType'
        locationId:
          type: string
          example: "LOC-001"
          nullable: true
        callIds:
          type: array
          items:
            type: string
          example: ["CALL-001", "CALL-002"]
          nullable: true
...
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'